app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
redis==4.5.5
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.14.0